pyproject.toml
[project]
name = "recipe-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-api-python-client>=2.188.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "google-genai>=1.61.0",
    "recipe-scrapers>=15.11.0",
]
